Refactor hammer_vlsi directory structure
