--- src/hammer-vlsi/hammer_vlsi.py
@@ -1,26 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-#
-#  hammer_vlsi.py
-#  Main entry point to the hammer_vlsi library.
-#
-#  Copyright 2018 Edward Wang <edward.c.wang@compdigitec.com>
-
-# Just import everything that the public hammer_vlsi module should see.
-
-import units
-
-from hammer_vlsi_impl import *
-# Remove imports unwanted in the public package
-from typing import List
-impl_remove = ["TimeValue", "VoltageValue", "TemperatureValue"]  # type: List[str]
-for name in impl_remove:
-    del globals()[name]
-
-from hammer_driver import *
-
-from verilog_utils import *
-
-from utils import *
-
-from cli_driver import CLIDriver
