+++ src/hammer-vlsi/hammer_vlsi/verilog_utils.py
@@ -8,6 +8,8 @@
 
 import re
 
+__all__ = ['VerilogUtils']
+
 
 class VerilogUtils:
     @staticmethod
